feat: create Balance atom
