+++ apps/dashboard/app/recoil/atoms/balanceAtom.ts
@@ -1,0 +1,11 @@
+import { atom } from "recoil";
+
+export const balanceState = atom({
+  key: "balanceState",
+  default: {
+    data: {
+      value: "0",
+    },
+    isLoading: true,
+  },
+});
